fixed shifting bug with integer literal

+++ board.cpp
@@ -1,5 +1,7 @@
 #include "board.hpp"
 
+#include <cstdint>
+
 Board::Board(Player* w, Player* b) {
     this->white = w;
     this->black = b;
@@ -9,6 +11,6 @@ void Board::updateBoardState() {
     this->boardstate = white->getBoardState() | black->getBoardState();
 }
 
-long long Board::getBoardState() {
+uint64_t Board::getBoardState() {
     return this->boardstate;
 }
