+++ moveQueen.hpp
@@ -2,10 +2,11 @@
 #define __MOVEQUEEN_HPP__
 
 #include "move.hpp"
+#include <cstdint>
 
 class MoveQueen: public Move {
     public:
-        bool testMove(long long position, long long newMove, long long playerState, long long boardState) const;
+        bool testMove(uint64_t position, uint64_t newMove, uint64_t playerState, uint64_t boardState) const;
 };
 
 #endif
